+++ modules/react/select/lib/Select.tsx
@@ -14,7 +14,6 @@ import {
   ExtractProps,
   createContainer,
   Themeable,
-  useLocalRef,
 } from '@workday/canvas-kit-react/common';
 import {system} from '@workday/canvas-tokens-web';
 
@@ -25,6 +24,11 @@ export interface SelectInputProps extends ExtractProps<typeof TextInput>, CSProp
    * ** Note:An option must be selected in order to render and icon.**
    */
   inputStartIcon?: CanvasSystemIcon;
+  readonly textInputProps?: {
+    ref: React.Ref<HTMLInputElement>;
+    onChange: () => {};
+    value: string;
+  };
 }
 
 const selectInputStencil = createStencil({
@@ -79,27 +83,6 @@ export const SelectInput = createSubcomponent(TextInput)({
     Element,
     model
   ) => {
-    const {localRef, elementRef} = useLocalRef(ref);
-
-    // We need to create a proxy between the multiple inputs. We need to redirect a few methods to
-    // the visible input
-    React.useImperativeHandle(
-      elementRef,
-      () => {
-        if (localRef.current) {
-          localRef.current.focus = (options?: FocusOptions) => {
-            textInputProps.ref.current!.focus(options);
-          };
-          localRef.current.blur = () => {
-            textInputProps.ref.current!.blur();
-          };
-        }
-
-        return localRef.current!;
-      },
-      [textInputProps.ref, localRef]
-    );
-
     return (
       <InputGroup data-width="ck-formfield-width">
         {inputStartIcon && model.state.selectedIds.length > 0 && (
@@ -117,7 +100,7 @@ export const SelectInput = createSubcomponent(TextInput)({
           onInput={onInput}
           value={value}
           name={name}
-          ref={elementRef}
+          ref={ref}
           {...hiddenSelectInputStencil()}
         />
         {/* Visual input */}